Add compute embeddings for options.
Embeddings are a numerical representation of text. They capture the meaning of the text. They are generated by a ML model (in our case bge-micro-v2) that was trained on lots of coherent text in english. The embedding model outputs an array of floats. We apply mean pooling on the result (what the model outputs) in order to always have the same array length on the result, no matter how long is the text that is processed (in our case the text that is processed is each option label). We also normalize the results in order for it to be ready for cosine similarity (will explain in a later commit). We'll always have the same embeddings for a given text calculated by same model and therefore we only need to calculate the embeddings for options once as long as the options array doesn't change. Basically when the component is loaded, we download from huggingface the ML model (https://huggingface.co/TaylorAI/bge-micro-v2/blob/main/onnx/model_quantized.onnx), which has around 17.4MB (note that the model will be taken from browser's cache after first download) and we calculate the embeddings for all options. Only then we display the options in UI since we don't have all data up to that point. Other notes (technical): We get an error if we try to send a function (getOptionLabel in our case) to the worker on the message, so we crete labelSemAutoCom property on the option object instead.