Add flow for userInput embeddings, compute cosine similarity and sort (sometimes filter too)
We disable the standard way of filtering by setting on filterOptions a function that returns what it receives. We do our own sorting and filtering when the input changes. Since the model is already loaded and the embeddings for the options are already calculated, each time the input changes, we only do the following: compute the embeddings for the current input text and then perform cosine similarity between each option embeddings and the input text embeddings in order to determine how similar a given option is to the input text. Note that if the option contains the input text (which is a standard search), we show that option with priority. We also filer if the component receives the threshold property (not a autcomplete property), which is the value of the cosine similarity above which we should display the options. Here is a nice video on embeddings: https://www.youtube.com/watch?v=gQddtTdmG_8